backend/requirements.txt
# Core API
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ML / Deep Learning
torch>=2.1.0
torchvision>=0.16.0
timm>=1.0.0

# Object Detection
ultralytics>=8.3.0

# Image Processing
opencv-python-headless>=4.9.0
Pillow>=10.0.0

# Data & Metrics
numpy>=1.24.0
scikit-learn>=1.3.0

# Dataset download
kagglehub>=0.3.0

# LLM (OpenRouter / Mistral)
requests>=2.31.0

# CORS
